refactor: move domain rules out of cpp_mcp_server.py
Extract three domain-rule concerns from the transport module into
focused modules:

- config_validation.py — _validate_config_file (absolute path, .json)
- query_policy.py — _parse_search_scope, _create_search_result,
  _check_tool_readiness (reads ctx.analyzer/state_manager)

This shrinks cpp_mcp_server.py to transport + dispatch + lifecycle
only. All handler and test imports updated to point at the new
modules. Zero behavioural change.
