+++ clang_index_mcp/_mcp/tool_handlers/call_graph_tools.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, List, Optional
 from mcp.types import TextContent
 
 from ..context import ctx
-from ..cpp_mcp_server import _create_search_result, _parse_search_scope
+from ..query_policy import _create_search_result, _parse_search_scope
 from ..response_formatters import suggestions
