+++ tests/test_search_scope.py
@@ -7,7 +7,7 @@ the old project_only boolean in the MCP tool interface.
 
 import pytest
 
-from clang_index_mcp._mcp.cpp_mcp_server import _VALID_SEARCH_SCOPES, _parse_search_scope
+from clang_index_mcp._mcp.query_policy import _VALID_SEARCH_SCOPES, _parse_search_scope
 
 
 class TestParseSearchScope:
